+++ Sources/SwiftUIKit/Documents/DocumentGroupOnboardingError.swift
@@ -1,9 +1,9 @@
 //
 //  DocumentGroupOnboardingError.swift
-//  OribiWriterKit
+//  SwiftUIKit
 //
 //  Created by Daniel Saidi on 2021-11-17.
-//  Copyright © 2021 Oribi. All rights reserved.
+//  Copyright © 2021 Daniel Saidi. All rights reserved.
 //
 
 import Foundation